+++ packages/userscript/src/parent/instance-manager.js
@@ -4,11 +4,7 @@ import { createInstanceManager as createSharedInstanceManager } from '@sremote/s
 export function createInstanceManager(customOptions = {}) {
   return createSharedInstanceManager({
     ns: NS,
-    logger: {
-      log: console_log,
-      debug: console_debug,
-      warn: console_warn,
-    },
+    logger: { log: console_log, debug: console_debug, warn: console_warn },
     onSignal: payload => {
       window.postMessage(payload, '*');
     },